Split TouchPatchBrowser into mixins; thin entry script.
TouchPatchBrowser now composes eight mixins in touch_browser_app.py;
touch_patch_browser.py remains the systemd entry point (14 lines).
